+++ changelog.md
@@ -8,6 +8,7 @@
   * Arrows shot to the dragon when at the center will now get sent back (without fire)
   * Ender Dragon now take 40% more damage from melee attacks when not at the center podium and 50% more damage when respawning crystals
   * Small pillars with cages now have obsidian blocks below the corner to prevent breaking them from below
+  * Crystals attached to the dragon when destroyed now deal 15% of max health as damage
   * Fixed Vanilla bugs
     * Fixed [MC-272431](https://bugs.mojang.com/browse/MC-272431)
       * This makes the ender dragon fly up and down faster, like she used to do back in 1.12